feat: implement enemy target manager
