+++ Assets/Scripts/EnemyManager/EnemyTargetManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d7d8ef1ade37f7c4ea1056c2269744e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
