+++ Assets/Scripts/Tower/TowerChildren/IceTower.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c68c281837acd0d48b373c1fe597bea5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
