+++ Assets/Scripts/TowerManager/TowerManager.cs
@@ -5,15 +5,16 @@ using UnityEngine;
 
 public partial class TowerManager // IO
 {
-	public void Launch(Tower obj) => _Launch(obj);
+	//public void Launch(Tower obj) => _Launch(obj);
 	public Bullet GetBullet(Tower tower) => _GetBullet(tower);
 	public void SetBullet(Bullet bullet) => _SetBullet(bullet);
-	public Enemy GetTarget() => _GetTarget();
+	public EnemyTargetManager GetTarget() => _GetTarget();
+	/*
 	public List<Enemy> GetNearTarget(float pin, float offset) => _GetNearTarget(pin, offset);
 
 	public Enemy GetNextTarget(float pin) => _GetNextTarget(pin);
 	public Enemy GetPrevTarget(float pin) => _GetPrevTarget(pin);
-
+*/
 	public bool AddTower() => _AddTower();
 	public void DestroyTower(Tower tower) => _DestroyTower(tower);
 
@@ -42,13 +43,12 @@ public partial class TowerManager // body
 	private readonly List<Tower> _towers = new List<Tower>();
 
 	// ReSharper disable Unity.PerformanceAnalysis
-	private void _Launch(Tower tower)
+	/*private void _Launch(Tower tower)
 	{
 		Bullet bullet = bulletPool.GetObject();
 		bullet.transform.position = tower.GetStartPosition();
-		bullet.SetTarget(GetTarget());
 		bullet.SetDamage(tower.towerData.damage * tower.GetGrade());
-	}
+	}*/
 
 	private Bullet _GetBullet(Tower tower)
 	{
@@ -73,11 +73,11 @@ public partial class TowerManager // body
 		return (true);
 	}
 
-	private Enemy _GetTarget()
+	private EnemyTargetManager _GetTarget()
 	{
-		return gameManager.enemyManager.targetFirst;
+		return gameManager.enemyManager.enemyTarget;
 	}
-
+/*
 	private List<Enemy> _GetNearTarget(float pin, float offset)
     {
 		return gameManager.enemyManager.GetNearEnemy(pin, offset);
@@ -92,7 +92,7 @@ public partial class TowerManager // body
     {
 		return gameManager.enemyManager.GetPrevTarget(pin);
     }
-
+*/
 	private void _Merge(Tower baseTower, Tower otherTower) {
 		if (baseTower.GetGrade() >= maxGrade)
 			return;
